# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: trips
    description: Trip management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips/batch:
    post:
      operationId: batchUpsertTrips
      tags: [trips]
      summary: Upsert multiple trips at once (for initial sync)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripBatch"
      responses:
        "200":
          description: Upserted trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips/{id}:
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    UserRegistration:
      type: object
      required: [email, name, plate, password]
      properties:
        email:
          type: string
        name:
          type: string
        plate:
          type: string
        password:
          type: string

    UserCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResult:
      type: object
      required: [token, email, name, plate]
      properties:
        token:
          type: string
        email:
          type: string
        name:
          type: string
        plate:
          type: string

    Trip:
      type: object
      required: [id, date, startAddr, endAddr, km, dur, type, deleted]
      properties:
        id:
          type: string
        date:
          type: string
        startAddr:
          type: string
        endAddr:
          type: string
        km:
          type: number
        dur:
          type: integer
        type:
          type: string
          enum: [business, private]
        edited:
          type: ["boolean", "null"]
        deleted:
          type: boolean

    TripInput:
      type: object
      required: [id, date, startAddr, endAddr, km, dur, type]
      properties:
        id:
          type: string
        date:
          type: string
        startAddr:
          type: string
        endAddr:
          type: string
        km:
          type: number
        dur:
          type: integer
        type:
          type: string
          enum: [business, private]
        edited:
          type: ["boolean", "null"]

    TripUpdate:
      type: object
      properties:
        date:
          type: string
        startAddr:
          type: string
        endAddr:
          type: string
        km:
          type: number
        dur:
          type: integer
        type:
          type: string
          enum: [business, private]
        edited:
          type: ["boolean", "null"]

    TripBatch:
      type: object
      required: [trips]
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/TripInput"