# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  token: string;
  email: string;
  name: string;
  plate: string;
}