# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripType } from "./tripType";

export interface Trip {
  id: string;
  date: string;
  startAddr: string;
  endAddr: string;
  km: number;
  dur: number;
  type: TripType;
  /** @nullable */
  edited?: boolean | null;
  deleted: boolean;
}