# lib/api-zod/src/generated/types/tripBatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripInput } from "./tripInput";

export interface TripBatch {
  trips: TripInput[];
}