# lib/api-zod/src/generated/types/tripInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripInputType } from "./tripInputType";

export interface TripInput {
  id: string;
  date: string;
  startAddr: string;
  endAddr: string;
  km: number;
  dur: number;
  type: TripInputType;
  /** @nullable */
  edited?: boolean | null;
}