# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripUpdateType } from "./tripUpdateType";

export interface TripUpdate {
  date?: string;
  startAddr?: string;
  endAddr?: string;
  km?: number;
  dur?: number;
  type?: TripUpdateType;
  /** @nullable */
  edited?: boolean | null;
}